requirements.txt
numpy
tensorflow
flwr
pandas
scikit-learn
matplotlib
seaborn
scipy
pyts
hdbscan
pyyaml
pydoc-markdown